+++ src/android/com/outsystems/plugins/sociallogins/AppleSignInActivity.kt
@@ -2,73 +2,57 @@ package com.outsystems.plugins.sociallogins
 
 import android.annotation.SuppressLint
 import android.app.Activity
-import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
-import androidx.browser.customtabs.CustomTabsIntent
-import androidx.browser.customtabs.CustomTabsClient
-
-import android.content.ComponentName
-
-import androidx.browser.customtabs.CustomTabsServiceConnection
-
-
-
+import android.webkit.WebView
 
 class AppleSignInActivity : Activity() {
     lateinit var redirectUri: String
     lateinit var clientId: String
+    lateinit var state: String
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        //setContentView(R.layout.activity_main)
-        setupAppleWebviewDialog(
-            "com.outsystems.mobile.plugin.sociallogin.apple",
-            "https://enmobile11-dev.outsystemsenterprise.com/SL_Core/rest/SocialLoginSignin/AuthRedirectOpenId")
-    }
 
-    override fun onNewIntent(intent: Intent?) {
-        super.onNewIntent(intent)
-        val code = intent?.data?.getQueryParameter("Code")
-        val user = intent?.data?.getQueryParameter("User")
-        val firstName = intent?.data?.getQueryParameter("firstName")
-        finish()
-    }
+        val bundle = intent.extras
+        if (bundle != null) {
 
-    @SuppressLint("SetJavaScriptEnabled", "NewApi")
-    fun setupAppleWebviewDialog(clientId: String, redirectUri: String/*, promise: Promise*/) {
-        this.clientId = clientId
-        this.redirectUri = redirectUri
+            if(intent.hasExtra("state") && intent.hasExtra("clientId") && intent.hasExtra("redirectUri")){
+                state = bundle.getString("state").toString()
+                clientId = bundle.getString("clientId").toString()
+                redirectUri = bundle.getString("redirectUri").toString()
 
-        val connection: CustomTabsServiceConnection = object : CustomTabsServiceConnection() {
-
-            override fun onCustomTabsServiceConnected(name: ComponentName, client: CustomTabsClient) {
-                openBrowser(client)
+                setupAppleWebviewDialog(
+                    clientId,
+                    redirectUri,
+                    state
+                )
             }
-
-            override fun onServiceDisconnected(name: ComponentName) {}
+            else{
+                setResult(11)
+                finish()
+            }
         }
-        val ok = CustomTabsClient.bindCustomTabsService(this, "com.android.chrome", connection)
-
     }
+    
+    @SuppressLint("SetJavaScriptEnabled", "NewApi")
+    fun setupAppleWebviewDialog(clientId: String, redirectUri: String, state: String) {
 
-    fun openBrowser(client : CustomTabsClient) {
+        //web view solution
+        val webView = WebView(this).apply {
+            settings.apply {
+                javaScriptEnabled = true
+                javaScriptCanOpenWindowsAutomatically = true
+            }
+        }
+        webView.isVerticalScrollBarEnabled = false
+        webView.isHorizontalScrollBarEnabled = false
+        webView.webViewClient = AppleWebViewClient(this)
 
-        val session = client.newSession(null)!!
 
-        //custom chrome tabs solution
-        val state = 1333
         val url: String = AppleConstants.AUTHURL + "?response_type=code id_token&response_mode=form_post&client_id=" + clientId + "&scope=" + AppleConstants.SCOPE + "&state=" + state + "&redirect_uri=" + redirectUri
 
-        val builder = CustomTabsIntent.Builder()
-        val customTabsIntent = builder
-            .setSession(session)
-            .build()
-
-        customTabsIntent.launchUrl(this, Uri.parse(url))
+        webView.loadUrl(url)
+        this.setContentView(webView)
     }
-
-
-
-
+    
 }