+++ src/android/com/outsystems/plugins/sociallogins/LoginDataResponse.kt
@@ -1,8 +1,0 @@
-package com.outsystems.plugins.sociallogins
-
-data class LoginDataResponse (
-    val id: String,
-    val email: String,
-    val name: String,
-    val photo: String
-)