+++ src/android/com/outsystems/plugins/sociallogins/UserInfo.kt
@@ -1,0 +1,10 @@
+package com.outsystems.plugins.sociallogins
+
+data class UserInfo (
+    val id: String?,
+    val email: String?,
+    val firstName: String?,
+    val lastName: String?,
+    val token: String?,
+    val picture: String? = ""
+)